+++ examples/resources/monad_alert_rule/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = monad_alert_rule.example
+  id = "c387bece-6a1d-4f2e-8b3c-9d0e1f2a3b4c"
+}