+++ examples/resources/monad_input/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = monad_input.example
+  id = "8f0c7a2e-4b1d-4c3e-9a6f-2d5e8b7c1a90"
+}