+++ examples/resources/monad_secret/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = monad_secret.example
+  id = "3f1c9e64-8a12-4c7d-9f30-1b2c4d5e6f70"
+}